+++ resources/js/components/search/SearchPostPopular.vue
@@ -633,6 +633,7 @@ export default {
       // 削除の確認のモーダル
       deletePostModalOpened: false,
       deletePostIndex: null,
+      deleteProssesing: false,
       // 投稿編集モーダル
       modalPostEditShow: false,
       editPostIndex: null,
@@ -694,25 +695,6 @@ export default {
           this.postsLoading = false;
         });
     },
-    // getPosts(word) {
-    //   // 読み込み中か最後のページなら読み込まない。
-    //   if (this.isLastPage) return;
-    //   if (this.itemLoading) return;
-    //   this.itemLoading = true;
-    //   axios.get('/api/post/search/popular/' + word + '?page=' + this.page)
-    //     .then((res) => {
-    //       console.log(res.data);
-    //       this.posts.push(...res.data.data);
-    //       this.itemLoading = false;
-    //       if (this.page === res.data.last_page) {
-    //         this.isLastPage = true;
-    //       }
-    //       this.page++;
-    //     }).catch((error) => {
-    //       console.log(error);
-    //       this.itemLoading = false;
-    //     });
-    // },
     // 無限スクロールのリセット
     resetPaginate() {
       this.posts = [];
@@ -766,12 +748,16 @@ export default {
     },
     // 投稿の削除
     deletePost() {
+      if (this.deleteProssesing) return;
+      this.deleteProssesing = true;
       axios.post('/api/post/delete/' + this.posts[this.deletePostIndex].id)
         .then(() => {
           this.posts.splice(this.deletePostIndex, 1);
           this.deletePostModalClose();
-        }).catch(() => {
-          return;
+          this.deleteProssesing = false;
+        }).catch((error) => {
+          console.log(error);
+          this.deleteProssesing = false;
         });
     },
     // 投稿編集モーダルを開く
@@ -826,7 +812,7 @@ export default {
       this.editErrors = [];
       this.nextNewImageId = -1;
       this.editHeight = '20px';
-      if (!this.edited) {
+      if (!this.editProcessing) {
         this.closePostMenu(this.editPostIndex);
       }
       this.modalPostEditShow = false;
@@ -908,9 +894,7 @@ export default {
           } else {
             this.posts.splice(this.editPostIndex, 1);
           }
-          this.edited = true;
           this.editPostModalClose();
-          this.edited = false;
           this.editProcessing = false;
         }).catch((error) => {
           this.editErrors = error.response.data.errors;
@@ -1042,7 +1026,10 @@ export default {
       axios.get('/api/donmai/users/' + this.modalPostId + '?page=' + this.donmaiPage)
         .then((res) => {
           console.log(res.data);
-          this.modalDonmaiUsers.push(...res.data.data);
+          const users = res.data.data.map((obj) => {
+            return obj.user;
+          });
+          this.modalDonmaiUsers.push(...users);
           this.donmaiLoading = false;
           if (this.donmaiPage === res.data.last_page) {
             this.isLastDonmaiPage = true;
@@ -1088,9 +1075,10 @@ export default {
       const img_height = img.height;
       if (img_height >= img_width) {
         this.heightIsBigger = true;
-        document.querySelector('.overlay-image-image').addEventListener('load', () => {
-          this.tatenagaImageWidth = document.querySelector('.overlay-image-image').clientWidth;
-          // console.log(this.tatenagaImageWidth);
+        this.$nextTick(function() {
+            this.tatenagaImageWidth = document.querySelector('.overlay-image-image').clientWidth;
+            this.handleResize();
+            // console.log(this.tatenagaImageWidth);
         });
       }
     },
@@ -1363,7 +1351,9 @@ export default {
       this.modalPostEditShow = false;
     }
     this.loadMorePosts = false;
-    next();
+    if (!this.editProcessing && !this.deleteProssesing) {
+      next();
+    }
   },
 
   beforeRouteUpdate (to, from, next) {
@@ -1377,7 +1367,9 @@ export default {
     }
     this.resetPaginate();
     this.getPosts(to.params.word);
-    next();
+    if (!this.editProcessing && !this.deleteProssesing) {
+      next();
+    }
   },
 }
 </script>
